[drape] Updated Visual Scale calculation.

+++ libs/map/framework.hpp
@@ -410,7 +410,7 @@ public:
   struct DrapeCreationParams
   {
     dp::ApiVersion m_apiVersion = dp::ApiVersion::OpenGLES3;
-    float m_visualScale = 1.0f;
+    double m_visualScale = 1.0;
     int m_surfaceWidth = 0;
     int m_surfaceHeight = 0;
     gui::TWidgetsInitInfo m_widgetsInitInfo;
